Module atsamd21g18a::i2s::clkctrl

Clock Unit n Control

Structs

FSINVR
Value of the field

FSOUTINVR
Value of the field

MCKDIVR
Value of the field

MCKENR
Value of the field

MCKOUTDIVR
Value of the field

MCKOUTINVR
Value of the field

NBSLOTSR
Value of the field

R
Value read from the register

SCKOUTINVR
Value of the field

W
Value to write to the register

_BITDELAYW
Proxy

_FSINVW
Proxy

_FSOUTINVW
Proxy

_FSSELW
Proxy

_FSWIDTHW
Proxy

_MCKDIVW
Proxy

_MCKENW
Proxy

_MCKOUTDIVW
Proxy

_MCKOUTINVW
Proxy

_MCKSELW
Proxy

_NBSLOTSW
Proxy

_SCKOUTINVW
Proxy

_SCKSELW
Proxy

_SLOTSIZEW
Proxy

Enums

BITDELAYR
Possible values of the field BITDELAY

BITDELAYW
Values that can be written to the field BITDELAY

FSSELR
Possible values of the field FSSEL

FSSELW
Values that can be written to the field FSSEL

FSWIDTHR
Possible values of the field FSWIDTH

FSWIDTHW
Values that can be written to the field FSWIDTH

MCKSELR
Possible values of the field MCKSEL

MCKSELW
Values that can be written to the field MCKSEL

SCKSELR
Possible values of the field SCKSEL

SCKSELW
Values that can be written to the field SCKSEL

SLOTSIZER
Possible values of the field SLOTSIZE

SLOTSIZEW
Values that can be written to the field SLOTSIZE