Module atsamd21g18a::tcc0::fctrlb

Recoverable Fault B Configuration

Structs

BLANKVALR
Value of the field

FILTERVALR
Value of the field

KEEPR
Value of the field

QUALR
Value of the field

R
Value read from the register

RESTARTR
Value of the field

W
Value to write to the register

_BLANKVALW
Proxy

_BLANKW
Proxy

_CAPTUREW
Proxy

_CHSELW
Proxy

_FILTERVALW
Proxy

_HALTW
Proxy

_KEEPW
Proxy

_QUALW
Proxy

_RESTARTW
Proxy

_SRCW
Proxy

Enums

BLANKR
Possible values of the field BLANK

BLANKW
Values that can be written to the field BLANK

CAPTURER
Possible values of the field CAPTURE

CAPTUREW
Values that can be written to the field CAPTURE

CHSELR
Possible values of the field CHSEL

CHSELW
Values that can be written to the field CHSEL

HALTR
Possible values of the field HALT

HALTW
Values that can be written to the field HALT

SRCR
Possible values of the field SRC

SRCW
Values that can be written to the field SRC