Module atsamd21g18a::tcc0::waveb

Waveform Control Buffer

Structs

CICCENB0R
Value of the field

CICCENB1R
Value of the field

CICCENB2R
Value of the field

CICCENB3R
Value of the field

CIPERENBR
Value of the field

POLB0R
Value of the field

POLB1R
Value of the field

POLB2R
Value of the field

POLB3R
Value of the field

R
Value read from the register

SWAPB0R
Value of the field

SWAPB1R
Value of the field

SWAPB2R
Value of the field

SWAPB3R
Value of the field

W
Value to write to the register

_CICCENB0W
Proxy

_CICCENB1W
Proxy

_CICCENB2W
Proxy

_CICCENB3W
Proxy

_CIPERENBW
Proxy

_POLB0W
Proxy

_POLB1W
Proxy

_POLB2W
Proxy

_POLB3W
Proxy

_RAMPBW
Proxy

_SWAPB0W
Proxy

_SWAPB1W
Proxy

_SWAPB2W
Proxy

_SWAPB3W
Proxy

_WAVEGENBW
Proxy

Enums

RAMPBR
Possible values of the field RAMPB

RAMPBW
Values that can be written to the field RAMPB

WAVEGENBR
Possible values of the field WAVEGENB

WAVEGENBW
Values that can be written to the field WAVEGENB